Groups v2

+++ servers/group/model/schedule.go
@@ -4,7 +4,7 @@ package model
 // It is used to determine the final meeting time
 type Schedule struct {
 	ScheduleID int    `json:"scheduleID"`
-	MeetingID  int    `json:"meetingID"`
+	MeetingID  int64  `json:"meetingID"`
 	StartTime  string `json:"startTime"`
 	EndTime    string `json:"endTime"`
 	Votes      int    `json:"votes"`
